Add context manager to turn off the interrupt signal handler.
There are plenty of places where we want to turn off the interrupt signal handler for a bit, and so it makes sense to rework that boilerplate into a context manager. Eventually, it might make sense to write it as a decorator, so that some functions run without signal interruption.